% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.r
\name{envOutliers}
\alias{envOutliers}
\title{Find outlying occurrence data in environmental space}
\usage{
envOutliers(
  pres,
  pvalSet = 1e-05,
  method = "grubbs",
  distEnvMethod = "euclidean",
  scaleData = TRUE,
  checkPairs = FALSE,
  kRosner = NULL
)
}
\arguments{
\item{pres}{An sf object (POINTS) or data.frame describing the environmental data
of species records.}

\item{pvalSet}{Numeric; user-specified p-value for assessing the significance of 
outlier test statistics. Default is 1e-5.}

\item{method}{Character; options are 'iqr', 'grubbs', 'dixon', 'rosner'. Default is 'grubbs'.}

\item{distEnvMethod}{Character; options are 'euclidian', 'manhattan', 'cosine,  or
'mahalanobis'. Default is 'euclidean'.}

\item{scaleData}{Boolean; Scale data using base::scale(). Default is TRUE.}

\item{checkPairs}{Logical; check for a single pair of outliers using the 
Grubbs test. This can only be performed for sample sizes <30. Only a single 
test is used because repeating it tends to throw out more points than seem 
reasonable, by eye. The value has no effect unless `method = 'grubbs'`. Default is FALSE.}

\item{kRosner}{Integer between 1 and 10. Determines the number of outliers 
suspected with a Rosner test. The value has no effect unless `method = 'rosner'`. 
Default is NULL.}
}
\value{
Returns the input sf object or data.frame with an additional column 'out_env' 
  indicating whether each point is considered a environmental outlier (TRUE) or not (FALSE).
}
\description{
Identifies environmental  outliers in species occurrence data.
}
\examples{
library(sf)
library(terra)
# Read data and convert to sf object
myPres <- read.csv(system.file('extdata/SpeciesCSVs/Camissonia_tanacetifolia.csv',
                               package='occOutliers'))
myPres <- myPres[complete.cases(myPres),]
myPres <- st_as_sf(myPres, coords = c("Longitude", "Latitude"))
# Load environmental data
env <- terra::rast(system.file('extdata/envs.tif', package = 'occOutliers'))
envData <- terra::extract(env, myPres, ID = FALSE)
myPres <- cbind(myPres, envData)
# Find environmental outliers
presOut <- envOutliers(pres = myPres, pvalSet = 1e-5)

}
\author{
Cory Merow <cory.merow@gmail.com>, Gonzalo E. Pinilla-Buitrago
}
