% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOutliers.r
\name{findOutlyingPoints}
\alias{findOutlyingPoints}
\title{Find outlying occurrence data}
\usage{
findOutlyingPoints(
  pres,
  spOutliers = TRUE,
  envOutliers = TRUE,
  method = "grubbs",
  pval = 1e-05,
  checkPairs = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{pres}{a `SpatialPoints` or `SpatialPointsDataFrame` object describing the locations of species records. A `SpatialPointsDataFrame` containing the values of environmental variables to be used must be supplied if `envOutliers=TRUE`}

\item{spOutliers}{logical; perform spatial outlier analysis}

\item{envOutliers}{logical; perform environmental outlier analysis}

\item{method}{character; options are 'iqr', 'grubbs', 'dixon', 'rosner'}

\item{pval}{user-specified p-value for assessing the significance of Grubbs test statistic.}

\item{checkPairs}{logical; check for a single pair of outliers. This can only be performed for sample sizes <30. Only a single test is used because repeating it tends to throw out more points than seem reasonable, by eye.}

\item{verbose}{logical; print messages}
}
\value{
Returns the SpatialPointsDataFrame provided with additional logical columns indicating spatial outliers (`spOutliers`) and environmental outliers ('`envOutliers`).
}
\description{
Spatial or environmental outliers
}
\details{
See Examples.
}
\examples{
myPres=read.csv(system.file('extdata/SpeciesCSVs/Camissonia_tanacetifolia.csv',
                            package='occOutliers'))
myPres=myPres[complete.cases(myPres),]
sp::coordinates(myPres)=c(1,2)
myEnv=raster::stack(system.file('extdata/AllEnv.tif',package='occOutliers'))
names(myEnv)=read.table(system.file('extdata/layerNames.csv',package='occOutliers'))[,1]
myPresDF=sp::SpatialPointsDataFrame(myPres,data.frame(raster::extract(myEnv,myPres)))
presOut=findOutlyingPoints(pres=myPresDF,
                           spOutliers=TRUE,
                           envOutliers=TRUE,
                           pval=1e-5)
world.shp=readRDS(system.file('extdata/worldShpMollwide.rds',package='occOutliers'))
plotOutliers(presOut,shpToPlot = world.shp)

}
\author{
Cory Merow <cory.merow@gmail.com>
}
