% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.r
\name{spatialOutliers}
\alias{spatialOutliers}
\title{Find outlying occurrence data in geographic space}
\usage{
spatialOutliers(
  pres,
  pvalSet = 1e-05,
  method = "grubbs",
  checkPairs = FALSE,
  kRosner = NULL
)
}
\arguments{
\item{pres}{An sf object (POINTS) describing the locations of species records. 
If environmental outliers are to be detected (not implemented in this function), 
the sf object should also contain the values of environmental variables to be used.}

\item{pvalSet}{Numeric; user-specified p-value for assessing the significance of 
outlier test statistics. Default is 1e-5.}

\item{method}{Character; options are 'iqr', 'grubbs', 'dixon', 'rosner'. Default is 'grubbs'.}

\item{checkPairs}{Logical; check for a single pair of outliers using the 
Grubbs test. This can only be performed for sample sizes <30. Only a single 
test is used because repeating it tends to throw out more points than seem 
reasonable, by eye. The value has no effect unless `method = 'grubbs'`. Default is FALSE.}

\item{kRosner}{Integer between 1 and 10. Determines the number of outliers 
suspected with a Rosner test. The value has no effect unless `method='rosner'`. 
Default is NULL.}
}
\value{
Returns the input sf object with an additional column 'out_spatial' 
  indicating whether each point is considered a spatial outlier (TRUE) or not (FALSE).
}
\description{
Identifies spatial outliers in species occurrence data.
}
\examples{
library(sf)

# Read data and convert to sf object
myPres <- read.csv(system.file('extdata/SpeciesCSVs/Camissonia_tanacetifolia.csv',
                               package='occOutliers'))
myPres <- myPres[complete.cases(myPres),]
myPres <- st_as_sf(myPres, coords = c("longitude", "latitude"), crs = 4326)

# Find spatial outliers
presOut <- spatialOutliers(pres = myPres, pvalSet = 1e-5)

}
\author{
Cory Merow <cory.merow@gmail.com>, Gonzalo E. Pinilla-Buitrago
}
