% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotOutliers}
\alias{plotOutliers}
\title{Plot outlying occurrence data}
\usage{
plotOutliers(
  pres,
  r = NULL,
  shpToPlot = NULL,
  legLoc = "topleft",
  bufferDist = 0.2
)
}
\arguments{
\item{pres}{an `sf` object of POINT geometry describing the locations of 
species records. It must contain columns for spatial and/or environmental 
outliers.}

\item{r}{an optional `terra::SpatRaster` object, representing environmental
data to plot in the background. If NULL, no environmental layer is plotted.}

\item{shpToPlot}{an optional `sf` object representing polygons to be 
plotted (e.g., a shapefile of boundaries or geographic areas). The shapefile 
is cropped to match the plotting extent of the presence points.}

\item{legLoc}{character, indicating the position of the legend on the plot. 
Options are `'bottomleft'`, `'topright'`, etc. If NULL, there is no legend.}

\item{bufferDist}{numeric, indicating the buffer distance as a proportion of 
the plot's largest axis (x or y) by which to expand the plot extent. Default 
is 0.5 (i.e., 50 percent of the largest axis).}
}
\description{
Plots spatial and/or environmental outliers for species 
 occurrence data.
}
\details{
This function plots the occurrence points, highlighting spatial and 
 environmental outliers. If environmental data or a shapefile is provided, it 
 is included in the background for context. The legend is dynamically 
 adjusted based on whether spatial and/or environmental outliers are detected.
}
\author{
Cory Merow <cory.merow@gmail.com>, Gonzalo E. Pinilla-Buitrago
}
